fix(openai): unexport ParseSSELine, remove duplicate v2/v3 streaming tests, extract shared test helpers
