+++ instrumentation/github.com/openai/openai-go/internal/streaming/streaming_test.go
@@ -326,7 +326,7 @@ func TestParseSSELine(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			payload, done := ParseSSELine(tt.line)
+			payload, done := parseSSELine(tt.line)
 			assert.Equal(t, tt.isDone, done)
 			if tt.payload != nil {
 				assert.Equal(t, tt.payload, payload)
